CR feedback updates: move array check to typeddataconverter
