+++ src/whylogs/core/columnprofile.py
@@ -1,8 +1,6 @@
 """
 Defines the ColumnProfile class for tracking per-column statistics
 """
-import numpy as np
-import pandas as pd
 
 from whylogs.core.statistics import (
     CountersTracker,
@@ -103,11 +101,11 @@ class ColumnProfile:
         Add `value` to tracking statistics.
         """
         self.counters.increment_count()
-        if ColumnProfile._are_nulls(value):
+        if TypedDataConverter._are_nulls(value):
             self.schema_tracker.track(InferredType.Type.NULL)
             return
 
-        if ColumnProfile._is_array_like(value):
+        if TypedDataConverter._is_array_like(value):
             return
 
         # TODO: ignore this if we already know the data type
@@ -116,7 +114,7 @@ class ColumnProfile:
         # TODO: Implement real typed data conversion
         typed_data = TypedDataConverter.convert(value)
 
-        if not ColumnProfile._are_nulls(typed_data):
+        if not TypedDataConverter._are_nulls(typed_data):
             self.cardinality_tracker.update(typed_data)
             self.frequent_items.update(typed_data)
         dtype = TypedDataConverter.get_type(typed_data)
@@ -130,17 +128,6 @@ class ColumnProfile:
         self.number_tracker.track(typed_data)
 
         self.constraints.update(typed_data)
-
-    @staticmethod
-    def _is_array_like(value):
-        return isinstance(value, list) or isinstance(value, np.ndarray) or isinstance(value, pd.Series)
-
-    @staticmethod
-    def _are_nulls(value):
-        if ColumnProfile._is_array_like(value):
-            return all(pd.isnull(value))
-        else:
-            return pd.isnull(value)
 
     def _unique_count_summary(self) -> UniqueCountSummary:
         cardinality_summary = self.cardinality_tracker.to_summary(_UNIQUE_COUNT_BOUNDS_STD)
